uploading new iframe
The iframe should be updated to this version in order to hide the unhelpful scrollbards from the page. Uploading a file so that we have the correct version documented here on Github, where it is also accessible to all of us. Feel free to comment and start a discussion about this, if you don't agree. I also opened an issue (see #5 ) about exactly this, where we can continue this discussion.